+++ lib/Extension/CompletionWorse/CompletionWorseExtension.php
@@ -6,6 +6,7 @@ use Closure;
 use Phpactor\Completion\Bridge\TolerantParser\LimitingCompletor;
 use Phpactor\Completion\Bridge\TolerantParser\ReferenceFinder\NameSearcherCompletor;
 use Phpactor\Completion\Bridge\TolerantParser\SourceCodeFilesystem\ScfClassCompletor;
+use Phpactor\Completion\Bridge\TolerantParser\TypeSuggestionProvider;
 use Phpactor\Completion\Bridge\TolerantParser\WorseReflection\DoctrineAnnotationCompletor;
 use Phpactor\Completion\Bridge\TolerantParser\WorseReflection\KeywordCompletor;
 use Phpactor\Completion\Bridge\TolerantParser\WorseReflection\ImportedNameCompletor;
@@ -13,6 +14,7 @@ use Phpactor\Completion\Bridge\TolerantParser\WorseReflection\WorseConstantCompl
 use Phpactor\Completion\Bridge\TolerantParser\WorseReflection\WorseConstructorCompletor;
 use Phpactor\Completion\Bridge\TolerantParser\WorseReflection\WorseDeclaredClassCompletor;
 use Phpactor\Completion\Bridge\TolerantParser\WorseReflection\WorseSignatureHelper;
+use Phpactor\Completion\Bridge\TolerantParser\WorseReflection\DocblockCompletor;
 use Phpactor\Completion\Bridge\WorseReflection\Formatter\ClassFormatter;
 use Phpactor\Completion\Bridge\WorseReflection\Formatter\ConstantFormatter;
 use Phpactor\Completion\Bridge\WorseReflection\Formatter\EnumCaseFormatter;
@@ -71,7 +73,7 @@ class CompletionWorseExtension implements Extension
     
     public function configure(Resolver $schema): void
     {
-        $completors = $this->getCompletors();
+        $completors = array_merge($this->getOtherCompletors(), $this->getTolerantCompletors());
         $defaults = array_combine(array_map(
             fn (string $key) => $this->completorEnabledKey($key),
             array_keys($completors)
@@ -92,7 +94,7 @@ class CompletionWorseExtension implements Extension
 
         $descriptions = array_combine(array_map(
             fn (string $key) => sprintf('completion_worse.completor.%s.enabled', $key),
-            array_keys($this->getCompletors())
+            array_keys($completors)
         ), array_map(
             fn (string $key, array $pair) => sprintf(
                 "Enable or disable the ``%s`` completor.\n\n%s.",
@@ -119,9 +121,14 @@ class CompletionWorseExtension implements Extension
 
     private function registerCompletion(ContainerBuilder $container): void
     {
-        foreach ($this->getCompletors() as $name => [$_, $completor]) {
+        foreach ($this->getTolerantCompletors() as $name => [$_, $completor]) {
             $container->register(sprintf('worse_completion.completor.%s', $name), $completor, [
                 self::TAG_TOLERANT_COMPLETOR => [ 'name' => $name ]
+            ]);
+        }
+        foreach ($this->getOtherCompletors() as $name => [$_, $completor]) {
+            $container->register(sprintf('worse_completion.completor.%s', $name), $completor, [
+                CompletionExtension::TAG_COMPLETOR => [ 'name' => $name ]
             ]);
         }
 
@@ -131,13 +138,6 @@ class CompletionWorseExtension implements Extension
                     return $container->get($serviceId);
                 }, $container->get(self::SERVICE_COMPLETOR_MAP)),
                 $container->get('worse_reflection.tolerant_parser')
-            );
-        }, [ CompletionExtension::TAG_COMPLETOR => []]);
-
-        $container->register(DoctrineAnnotationCompletor::class, function (Container $container) {
-            return new DoctrineAnnotationCompletor(
-                $container->get(NameSearcher::class),
-                $container->get(WorseReflectionExtension::SERVICE_REFLECTOR)
             );
         }, [ CompletionExtension::TAG_COMPLETOR => []]);
 
@@ -232,11 +232,28 @@ class CompletionWorseExtension implements Extension
             [ CompletionExtension::TAG_SNIPPET_FORMATTER => []]
         );
     }
+    /**
+     * @return array<string,array{string,Closure(Container): mixed}>
+     */
+    private function getOtherCompletors(): array
+    {
+        return [
+            'doctrine_annotation' => [
+                'Completion for annotations provided by the Doctrine annotation library',
+                function (Container $container) {
+                    return new DoctrineAnnotationCompletor(
+                        $container->get(NameSearcher::class),
+                        $container->get(WorseReflectionExtension::SERVICE_REFLECTOR)
+                    );
+                },
+            ],
+        ];
+    }
 
     /**
      * @return array<string,array{string,Closure(Container): mixed}>
      */
-    private function getCompletors(): array
+    private function getTolerantCompletors(): array
     {
         return [
             'imported_names' => [
@@ -342,6 +359,15 @@ class CompletionWorseExtension implements Extension
                 'Completion for keywords (not very accurate)',
                 function (Container $container) {
                     return new KeywordCompletor();
+                },
+            ],
+            'docblock' => [
+                'Docblock completion',
+                function (Container $container) {
+                    return new DocblockCompletor(
+                        new TypeSuggestionProvider($container->get(NameSearcher::class)),
+                        $container->get(WorseReflectionExtension::SERVICE_PARSER)
+                    );
                 },
             ],
         ];