+++ lib/Extension/LanguageServerCompletion/Handler/CompletionHandler.php
@@ -124,7 +124,7 @@ class CompletionHandler implements Handler, CanRegisterCapabilities
 
     public function registerCapabiltiies(ServerCapabilities $capabilities): void
     {
-        $capabilities->completionProvider = new CompletionOptions([':', '>', '$']);
+        $capabilities->completionProvider = new CompletionOptions([':', '>', '$', '@']);
         $capabilities->signatureHelpProvider = new SignatureHelpOptions(['(', ',']);
     }