+++ CHANGELOG.md
@@ -59,6 +59,7 @@ Improvements:
 
 Features:
 
+  - [compl] Docblock completion
   - [rf] Union type support for goto definition
   - [ct] Add missing PHP return types
   - [wr] Support for inference for `array_map`, and arrow and anonymous functions